Fix unsigned 32-bit overflow in out_contribs pointer arithmetic
When num_samples * (num_features + 1) * num_outputs exceeds 2^32, the
unsigned arithmetic in out_contribs offset computation silently wraps
around, causing SHAP values to be written to wrong memory locations.
This produces garbage results (e.g. mean additivity mismatch of 0.178
with 8M samples) without any error or warning.

Cast the loop index to unsigned long long before multiplication in all
five out_contribs offset sites: dense_tree_saabas, tree_shap (main),
dense_independent, dense_tree_path_dependent (already fixed), and
dense_tree_interactions_path_dependent.

Fixes #4151
Fixes #4002

+++ shap/cext/tree_shap.h
@@ -326,7 +326,7 @@ inline void dense_tree_saabas(tfloat *out_contribs, const TreeEnsemble& trees, c
 
     // build explanation for each sample
     for (unsigned i = 0; i < data.num_X; ++i) {
-        instance_out_contribs = out_contribs + i * (data.M + 1) * trees.num_outputs;
+        instance_out_contribs = out_contribs + static_cast<unsigned long long>(i) * (data.M + 1) * trees.num_outputs;
         data.get_x_instance(instance, i);
 
         // aggregate the effect of explaining each tree
@@ -1201,7 +1201,7 @@ inline void dense_independent(const TreeEnsemble& trees, const ExplanationDatase
         for (unsigned i = 0; i < data.num_X; ++i) {
             const tfloat *x = data.X + i * data.M;
             const bool *x_missing = data.X_missing + i * data.M;
-            instance_out_contribs = out_contribs + i * (data.M + 1) * trees.num_outputs;
+            instance_out_contribs = out_contribs + static_cast<unsigned long long>(i) * (data.M + 1) * trees.num_outputs;
             const tfloat y_i = data.y == NULL ? 0 : data.y[i];
 
             print_progress_bar(last_print, start_time, oind * data.num_X + i, data.num_X * trees.num_outputs);
@@ -1356,7 +1356,7 @@ inline void dense_tree_interactions_path_dependent(const TreeEnsemble& trees, co
     tfloat *on_contribs = new tfloat[contrib_row_size];
     tfloat *off_contribs = new tfloat[contrib_row_size];
     for (unsigned i = 0; i < data.num_X; ++i) {
-        instance_out_contribs = out_contribs + i * (data.M + 1) * contrib_row_size;
+        instance_out_contribs = out_contribs + static_cast<unsigned long long>(i) * (data.M + 1) * contrib_row_size;
         data.get_x_instance(instance, i);
 
         // aggregate the effect of explaining each tree
@@ -1432,7 +1432,7 @@ inline void dense_global_path_dependent(const TreeEnsemble& trees, const Explana
     ExplanationDataset instance;
     tfloat *instance_out_contribs;
     for (unsigned i = 0; i < data.num_X; ++i) {
-        instance_out_contribs = out_contribs + i * (data.M + 1) * trees.num_outputs;
+        instance_out_contribs = out_contribs + static_cast<unsigned long long>(i) * (data.M + 1) * trees.num_outputs;
         data.get_x_instance(instance, i);
 
         // since we now just have a single merged tree we can just use the tree_path_dependent algorithm
